EDU-196 Move mapping logic to client data mapper
