+++ src/server/room-controller.js
@@ -6,14 +6,16 @@ import RoomService from '../services/room-service.js';
 import needsPermission from '../domain/needs-permission-middleware.js';
 import { validateBody, validateParams } from '../domain/validation-middleware.js';
 import { roomDetailsParamSchema, roomSchema } from '../domain/schemas/rooms-schemas.js';
+import ClientDataMapper from './client-data-mapper.js';
 
 const { NotFound } = httpErrors;
 export default class RoomController {
-  static get inject() { return [RoomService, PageRenderer]; }
+  static get inject() { return [RoomService, ClientDataMapper, PageRenderer]; }
 
-  constructor(roomService, pageRenderer) {
+  constructor(roomService, clientDataMapper, pageRenderer) {
     this.roomService = roomService;
     this.pageRenderer = pageRenderer;
+    this.clientDataMapper = clientDataMapper;
   }
 
   registerApi(router) {
@@ -26,11 +28,13 @@ export default class RoomController {
   registerPages(router) {
     router.get('/rooms/:roomId', [needsPermission(permissions.VIEW_ROOMS), validateParams(roomDetailsParamSchema)], async (req, res) => {
       const { roomId } = req.params;
-      const roomDetails = await this.roomService.getRoomDetailsById(roomId);
+      const room = await this.roomService.getRoomById(roomId);
 
-      if (!roomDetails) {
+      if (!room) {
         throw new NotFound();
       }
+
+      const roomDetails = await this.clientDataMapper.mapRoomDetails(room);
 
       return this.pageRenderer.sendPage(req, res, PAGE_NAME.room, { roomDetails });
     });
