+++ src/services/room-service.js
@@ -21,31 +21,7 @@ export default class RoomService {
     return newRoom;
   }
 
-  async getRoomDetailsById(roomId) {
-    const roomDetails = await this.roomsStore.findOne({ _id: roomId });
-
-    if (!roomDetails) {
-      return null;
-    }
-
-    const owner = await this.userService.getUserById(roomDetails.owner);
-    roomDetails.owner = {
-      userId: owner._id,
-      username: owner.username,
-      email: owner.email
-    };
-
-    const members = await this.userService.getUsersByIds(roomDetails.members.map(member => member.userId));
-
-    roomDetails.members = roomDetails.members.map(member => {
-      const memberDetails = members.find(user => member.userId === user._id);
-      return {
-        ...member,
-        email: memberDetails.email,
-        username: memberDetails.username
-      };
-    });
-
-    return roomDetails;
+  getRoomById(roomId) {
+    return this.roomsStore.findOne({ _id: roomId });
   }
 }
